Split v2 eval tests module
